+++ src/formatters/map.ts
@@ -3,9 +3,7 @@ import util from '../util'
 import { represent } from '../common'
 
 export function formatMap(this: Map<unknown, unknown>, fmt: Formatter) {
-    const name = Reflect.getPrototypeOf(this).constructor.name
-
-    fmt.map(name, fmt => {
+    fmt.map(this, fmt => {
         for (const [key, value] of this) {
             fmt.entry(key, value)
         }
@@ -14,7 +12,6 @@ export function formatMap(this: Map<unknown, unknown>, fmt: Formatter) {
 util.extend(Map, represent, formatMap)
 
 export function formatWeakMap(this: WeakMap<object, unknown>, fmt: Formatter) {
-    const name = Reflect.getPrototypeOf(this).constructor.name
-    fmt.write(name)
+    fmt.write(util.objectName(this)!)
 }
 util.extend(WeakMap, represent, formatWeakMap)