+++ src/formatters/object.ts
@@ -3,11 +3,7 @@ import util from '../util'
 import { represent } from '../common'
 
 export function formatObject(this: object, fmt: Formatter) {
-    const prototype = Reflect.getPrototypeOf(this)
-    const constructor = prototype && prototype.constructor
-    const name = constructor === Object || !constructor ? null : constructor.name
-
-    fmt.struct(name, fmt => {
+    fmt.struct(this, fmt => {
         for (const key of Reflect.ownKeys(this).sort(util.compareKeys)) {
             if (Reflect.getOwnPropertyDescriptor(this, key)!.enumerable) {
                 fmt.field(key, this[key as keyof object])