+++ src/formatters/set.ts
@@ -3,9 +3,7 @@ import util from '../util'
 import { represent } from '../common'
 
 export function formatSet(this: Set<unknown>, fmt: Formatter) {
-    const name = Reflect.getPrototypeOf(this).constructor.name
-
-    fmt.set(name, fmt => {
+    fmt.set(this, fmt => {
         for (const value of this) {
             fmt.entry(value)
         }
@@ -14,7 +12,6 @@ export function formatSet(this: Set<unknown>, fmt: Formatter) {
 util.extend(Set, represent, formatSet)
 
 export function formatWeakSet(this: WeakSet<object>, fmt: Formatter) {
-    const name = Reflect.getPrototypeOf(this).constructor.name
-    fmt.write(name)
+    fmt.write(util.objectName(this)!)
 }
 util.extend(WeakSet, represent, formatWeakSet)