+++ src/util/index.ts
@@ -2,10 +2,12 @@ import compareKeys from './compareKeys'
 import escape from './escape'
 import extend from './extend'
 import isIdentifier from './isIdentifier'
+import objectName from './objectName'
 
 export default {
     compareKeys,
     escape,
     extend,
     isIdentifier,
+    objectName,
 }